+++ NEWS.md
@@ -1,9 +1,16 @@
+# dtq 0.1.9.3
+
+- [ ] add source name to `dtl(chain=TRUE)`
+- [ ] more unit tests
+- [x] removed internal not currently used functions
+
+
 # dtq 0.1.9.2
 
 - [x] dtq class new methods
 - [x] vectorizing a lot of functions
 - [x] log source data.table name
-- [ ] more unit tests
+- [x] more unit tests
 
 
 # dtq 0.1.9